Add test_util library and tests/ directory for unit testing
Move include/ src/ to separate CMake directory - image/, similar to
the tests/ and test_util/ directories.

Currently the test_util library provides two testing macros: ASSERT_EQ and ASSERT_NEQ
Use for checking for (in)equality. The library also includes a global operator<<
for iostream formatting, which uses SFINAE to target only types which haven't
overloaded this operator. The behvior is to print the address and type name
of the operand by means of typeid(operand).name. With GCC targets an attempt
is made to demangle the mangled name returned by typeid(..).name.

The new layout aims to be as modular - the test_util code is not linked
to the main library - it is used only by unit test executables.

Also enable compiler warnings for GCC and MSVC.